Create Counting Algorithm
Working on making an algorithm prototype which will analyse the sound wave to see if it is above our target amplitude enough to register as a "count".

In place of sensor data I will have a random number generator. I imagine the real thing will be an output array to plot the range of frequencies? In that case the implementation would have to be some kind of array analysis to see if the portion of the array plotting our target frequency (~4750-5250 Hz) has a majority of its values above that target. 
Just my prediction, I'm still unsure how that microphone output is wrapped to be plotted